refactor(int32): philo id is not int32, refactored a bunch of functions and structs as a result
This comes as a result of no longuer considering argc as a size_t when passing
it to initialisation functions.

+++ philo/include/message_bus_send.h
@@ -5,19 +5,18 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By:  <>                                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/04/11 19:05:22 by                   #+#    #+#             */
-/*   Updated: 2023/04/11 19:22:51 by emcnab           ###   ########.fr       */
+/*   Created: 2023/04/13 11:26:42 by                   #+#    #+#             */
+/*   Updated: 2023/04/13 11:42:08 by emcnab           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MESSAGE_BUS_SEND_H
 # define MESSAGE_BUS_SEND_H
 
-# include <stddef.h>
 # include <stdint.h>
 
 # include "e_philo_state.h"
 
-int32_t	message_bus_send(size_t id, t_e_philo_state state);
+int32_t	message_bus_send(int32_t id, t_e_philo_state state);
 
 #endif
