+++ philo/include/s_philo.h
@@ -6,7 +6,7 @@
 /*   By: emcnab <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/06 18:08:42 by emcnab            #+#    #+#             */
-/*   Updated: 2023/04/11 19:04:14 by emcnab           ###   ########.fr       */
+/*   Updated: 2023/04/13 11:15:41 by emcnab           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include <pthread.h>
 # include <stddef.h>
+# include <stdint.h>
 
 # include "e_philo_state.h"
 # include "s_promise.h"
@@ -22,7 +23,7 @@
 
 typedef struct s_philo
 {
-	size_t			id;
+	int32_t			id;
 	t_e_philo_state	state;
 	pthread_mutex_t	fork_left;
 	pthread_mutex_t	fork_right;
