+++ philo/include/table_get_right.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   table_get_right.h                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By:  <>                                        +#+  +:+       +#+        */
+/*   By: emcnab <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/04/10 17:56:11 by                   #+#    #+#             */
-/*   Updated: 2023/04/10 17:56:27 by emcnab           ###   ########.fr       */
+/*   Created: 2023/04/13 11:10:44 by emcnab            #+#    #+#             */
+/*   Updated: 2023/04/13 11:10:47 by emcnab           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@
 # include "s_philo.h"
 # include "s_table.h"
 
-t_s_philo	*tables_get_right(t_s_table *table, size_t seat);
+t_s_philo	*tables_get_right(t_s_table *table, int32_t seat);
 
 #endif
